Move all header buttons to menu modal if screen size is small
