+++ src/client/app/components/HeaderComponent.jsx
@@ -4,107 +4,38 @@
 
 import React from 'react';
 import { Link } from 'react-router';
-import { Button } from 'reactstrap';
 import LogoComponent from './LogoComponent';
-import UIModalComponent from './UIModalComponent';
-import { hasToken } from '../utils/token';
+import HeaderButtonsComponent from './HeaderButtonsComponent';
 
 /**
  * React component that controls the header strip at the top of all pages
- * @param props The props passed down by the parent component
- * @return JSX to create the header strip
  */
-export default class HeaderComponent extends React.Component {
-	constructor(props) {
-		super(props);
-		this.handleLogOut = this.handleLogOut.bind(this);
-	}
+export default function HeaderComponent(props) {
+	const titleStyle = {
+		marginTop: '5px',
+		display: 'inline-block'
+	};
+	const divStyle = {
+		paddingBottom: '5px'
+	};
+	const divRightStyle = {
+		marginTop: '5px',
+		display: 'flex'
+	};
 
-	handleLogOut() {
-		localStorage.removeItem('token');
-		this.forceUpdate();
-	}
-
-	render() {
-		const urlArr = window.location.href.split('/');
-		const page = urlArr[urlArr.length - 1];
-		let renderOptionsButton = false;
-		let renderLoginButton = false;
-		let renderAdminButton = false;
-		let renderGroupsButton = false;
-		const renderLogoutButton = hasToken();
-
-		switch (page) {
-			case '': // home page
-				renderOptionsButton = true;
-				if (renderLogoutButton) {
-					renderAdminButton = true;
-					renderGroupsButton = true;
-				} else {
-					renderLoginButton = true;
-				}
-				break;
-			case 'groups':
-				renderAdminButton = true;
-				break;
-			case 'admin':
-				renderGroupsButton = true;
-				break;
-			case 'login':
-				break;
-			default: // Unknown page, routes to 404, show nothing
-				break;
-
-		}
-
-		const titleStyle = {
-			marginTop: '5px',
-			display: 'inline-block'
-		};
-		const divStyle = {
-			paddingBottom: '5px'
-		};
-		const divRightStyle = {
-			marginTop: '5px',
-			display: 'flex'
-		};
-		const loginLinkStyle = {
-			display: renderLoginButton ? 'inline' : 'none',
-			paddingLeft: '5px'
-		};
-		const adminLinkStyle = {
-			display: renderAdminButton ? 'inline' : 'none',
-			paddingLeft: '5px'
-		};
-		const groupsLinkStyle = {
-			display: renderGroupsButton ? 'inline' : 'none',
-			paddingLeft: '5px'
-		};
-		const logoutButtonStyle = {
-			display: renderLogoutButton ? 'inline' : 'none',
-			paddingLeft: '5px'
-		};
-
-		return (
-			<div className="container-fluid" style={divStyle}>
-				<div className="row">
-					<div className="col-4">
-						<Link to="/"><LogoComponent url="./app/images/logo.png" /></Link>
-					</div>
-					<div className="col-4 text-center">
-						<h1 style={titleStyle}>{this.props.title}</h1>
-					</div>
-					<div className="col-4 justify-content-end" style={divRightStyle}>
-						<div className="d-lg-none">
-							{(renderOptionsButton) ? <UIModalComponent /> : null}
-						</div>
-						<Link style={loginLinkStyle} to="/login"><Button outline>Log In</Button></Link>
-						<Link style={adminLinkStyle} to="/admin"><Button outline>Admin panel</Button></Link>
-						<Link style={groupsLinkStyle} to="/groups"><Button outline>Groups</Button></Link>
-						<Link style={logoutButtonStyle} to="/"><Button outline onClick={this.handleLogOut}>Log Out</Button></Link>
-					</div>
+	return (
+		<div className="container-fluid" style={divStyle}>
+			<div className="row">
+				<div className="col-4">
+					<Link to="/"><LogoComponent url="./app/images/logo.png" /></Link>
+				</div>
+				<div className="col-4 text-center">
+					<h1 style={titleStyle}>{props.title}</h1>
+				</div>
+				<div className="col-4 justify-content-end" style={divRightStyle}>
+					<HeaderButtonsComponent renderOptionsButton />
 				</div>
 			</div>
-		);
-	}
+		</div>
+	);
 }
